+++ Runway_Redeclaration/src/DeclaredRunwayInterface.java
@@ -35,27 +35,27 @@ public interface DeclaredRunwayInterface {
 	/**
 	 * Take-Off-Runway-Available
 	 */
-	int getTORA();
+	double getTORA();
 	
-	int getClearway();
+	double getClearway();
 	
-	int getStopway();
+	double getStopway();
 	
-	int getDisplacedThreshold();
+	double getDisplacedThreshold();
 	
 	/**
 	 * ASDA = TORA + Stopway
 	 */
-	int getASDA();
+	double getASDA();
 	
 	/**
 	 * TODA = TORA + Clearway
 	 */
-	int getTODA();
+	double getTODA();
 	
 	/**
 	 * LDA = TORA - Displaced Threshold
 	 */
-	int getDLA();
+	double getDLA();
 	
 }
