+++ Runway_Redeclaration/src/Obstacle.java
@@ -6,7 +6,7 @@
  */
 public class Obstacle implements ObstacleInterface {
 	private String name;
-	private int radius, height;
+	private double radius, height;
 	
 	public Obstacle(String name, int farthestPoint, int highestPoint){
 		this.name = name;
@@ -20,12 +20,12 @@ public class Obstacle implements ObstacleInterface {
 	}
 
 	@Override
-	public int getHeight() {
+	public double getHeight() {
 		return this.height;
 	}
 
 	@Override
-	public int getRadius() {
+	public double getRadius() {
 		return this.radius;
 	}
